Upgrade gems
This upgrades the ruby version and the gems to the latest. One thing to note is that there is one test failing. It seems as though paperclip isn't saving files with messages anymore. I could spend more time fixing this, but the very next thing to do is to remove paperclip, so I don't feel it's worth the time.

Let's get this merged and move forward with our paperclip alternative.