Base64 file contents and cleanup the controller
A lot of the work here came from refactoring the controller. But I also needed to base64 encode the file contents. The Message model could probably use a little refactoring, and we want to ensure that the contents do get encoded since they are decoded on read, but this does clean things up a lot at the moment.